Doing the junk and stuff needed to get the block to actually work on Calypso
